requirements.txt
asgiref==3.6.0
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
crashtest==0.4.1
cryptography==39.0.1
distlib==0.3.6
Django==4.1.7
dulwich==0.21.3
filelock==3.9.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.4
packaging==23.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry-core==1.5.1
poetry-plugin-export==1.3.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.2.20
urllib3==1.26.14
virtualenv==20.20.0
webencodings==0.5.1
zipp==3.15.0
djangorestframework==3.14.0
django-filter==23.1
drf-spectacular==0.26.0
inflection==0.5.1
uritemplate==4.1.1
defusedxml==0.7.1
python3-openid==3.2.0
social-auth-app-django==5.1.0
social-auth-core==4.4.0

